+++ SpringGame_2026/SpringGame_2026/Scene/TitleScene.cpp
@@ -172,7 +172,7 @@ void TitleScene::Draw()
 	//画面下部に「Press A Button」の文字を描画する
 	std::string startText = "Press A Button";
 	//描画する文字列の横幅を取得する
-	int textWidth = GetDrawStringWidthToHandle(startText.c_str(), startText.length(), m_startFontHandle);
+	int textWidth = GetDrawStringWidthToHandle(startText.c_str(), static_cast<int>(startText.length()), m_startFontHandle);
 	//画面下部の中心に描画するため、描画位置を計算する
 	Vector3 textPos = { windowSize.w / 2.0f - textWidth / 2.0f, windowSize.h * title_logo_pos_y_rate, 0.0f };
 	//描画する
